projects/py_projects/HoB/requirements.txt
tox
PySocks
python-telegram-bot
sqlalchemy
pylint
mypy_extensions
